+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 * [#604](https://github.com/clojure-emacs/clojure-mode/pull/604): Add `bb` (babashka) to `interpreter-mode-alist`.
 
+### Bugs fixed
+
+* [#608](https://github.com/clojure-emacs/clojure-mode/issues/608) Fix alignment issue involving margin comments at the end of nested forms.
+
 ## 5.13.0 (2021-05-05)
 
 ### New features
@@ -26,7 +30,6 @@
 
 * [#588](https://github.com/clojure-emacs/clojure-mode/pull/588): Fix font-lock for character literals.
 * Stop `clojure-sort-ns` from calling `redisplay`.
-* [#608](https://github.com/clojure-emacs/clojure-mode/issues/608) Fix alignment issue involving margin comments at the end of nested forms.
 
 ### Changes