+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+* [#622](https://github.com/clojure-emacs/clojure-mode/issues/622): Add font locking for missing clojure.core macros
 * [#615](https://github.com/clojure-emacs/clojure-mode/issues/615): Support clojure-dart files.
 
 ## 5.14.0 (2022-03-07)